Attach Bearer token authorization in AuthInterceptor; Moved TokenManager to util package
